requirements.txt
asgiref==3.4.1
Babel==2.9.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.24
django-allauth==0.45.0
django-babel==0.6.2
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.8
idna==3.2
oauthlib==3.1.1
phonenumbers==8.12.28
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
